% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema_db.R
\name{create_schema_db}
\alias{create_schema_db}
\title{Create UK Biobank data schema database}
\usage{
create_schema_db(file = "", path = ".", date = Sys.Date(),
  silent = !interactive(), debug = FALSE, overwrite = FALSE,
  as_is = FALSE)
}
\arguments{
\item{file}{The filename for the schema database. Defaults to \code{""}, which is
interpreted as \code{paste0("ukb-schema-", date, ".sqlite")}. If this file already
exists in directory \code{path}, the session is interactive, and \code{overwrite} is
not \code{FALSE}, then the user will be prompted to decide whether the file should
be overwritten.}

\item{path}{The path to the directory where the file will be saved. Defaults
to \code{.} (the current directory).}

\item{date}{The date-stamp for the default filename. Defaults to the current
date in \code{YYYY-MM-DD} format.}

\item{silent}{Do not report progress. Defaults to \code{FALSE}.}

\item{debug}{Report debugging information (useful when e.g. the structure of
the schema files has changed). Defaults to \code{FALSE}.}

\item{overwrite}{Always overwrite existing files? Helpful for non-interactive
use. Defaults to \code{FALSE}.}

\item{as_is}{Import the schemas into the database without tidying? Defaults
to \code{FALSE}.}
}
\value{
A database connection object of class
\link[RSQLite:SQLiteConnection-class]{RSQLite::SQLiteConnection}
}
\description{
\code{create_schema_db()} generates an SQLite database containing the UK Biobank
data dictionaries from http://biobank.ctsu.ox.ac.uk/crystal/schema.cgi
}
\details{
Most of the workings of \code{create_schema_db()} are hidden from the
user; the function downloads all the schema files and loads them into
an SQLite database with pre-defined table structure. Note that if the
table structure has changed (i.e. has been changed by UK Biobank), then the
function will fail partially or fully. Debugging information may be helpful
to diagnose and/or fix such failures. Fixes currently require changes to
the \code{.sql} files in the \code{sql/} subsdirectory of the package installation.
}
