% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-db.R
\name{load_db}
\alias{load_db}
\title{Load schemas from a ukbschemas database}
\usage{
load_db(file = NULL, db = NULL)
}
\arguments{
\item{file}{Full path to an SQLite database, or \code{NULL}.}

\item{db}{A (possibly disconnected) database connection, or \code{NULL}.}
}
\value{
A named list with elements of class \code{tbl_df}, containing the tables
from \code{db}.
}
\description{
\code{load_db()} is a simple front-end to \link{DBI} functions which can be
used to load tables from a ukbschemas database.
}
\details{
\code{load_db()} will attempt to open \code{file} as an SQLite database,
or will attempt to open connection \code{db} (if closed) and will throw errors if
the file is not a valid databse or it is impossible to create a valid
connection from \code{db}. Tables are read with \link[DBI:dbReadTable]{DBI::dbReadTable} and converted
to data frames of class \code{tbl_df} (see \link[tibble:tibble]{tibble::tibble}). The database
connection is always closed before the function returns its result.
}
\examples{
\dontrun{
db <- ukbschemas_db(path = tempdir())
sch <- load_db(db = db)
}
}
