% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_schemas.R
\name{load_schemas}
\alias{load_schemas}
\title{Load schemas from a ukbschema database}
\usage{
load_schemas(db)
}
\arguments{
\item{db}{A (possibly disconnected) database connection.}
}
\value{
A named list with elements of class \link[tibble:tbl_df]{tibble::tbl_df}, containing the
tables from \code{db}.
}
\description{
\code{load_schemas()} is a simple front-end to \link[DBI:DBI]{DBI::DBI} functions which can be
used to load tables from a ukbschema database.
}
\details{
\code{load_schemas()} will attempt to open connection \code{db} and will
throw an error if this is not possible. Tables are read with
\link[DBI:dbReadTable]{DBI::dbReadTable} and converted to data frames of class \code{tbl_df}. The
database connection is always closed after the tables are loaded.
}
