% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-db.R
\name{save_db}
\alias{save_db}
\title{Save a list of UK Biobank data schemas to an SQLite database}
\usage{
save_db(
  sch,
  file = "",
  path = ".",
  date_str = Sys.Date(),
  silent = !interactive(),
  overwrite = FALSE,
  as_is = FALSE
)
}
\arguments{
\item{sch}{List of data frames, representing UK Biobank data schemas (unless \code{as_is == TRUE}, in
which case any list of data frames is permitted).}

\item{file}{The filename for the schema database. Defaults to \code{""}, which is interpreted as
\code{paste0("ukb-schemas-", date, ".sqlite")}. If this file already exists in directory \code{path}, the
session is interactive, and \code{overwrite} is not \code{FALSE}, then the user will be prompted to decide
whether the file should be overwritten.}

\item{path}{The path to the directory where the file will be saved. Defaults to \code{.} (the current
directory).}

\item{date_str}{The date-stamp for the default filename. Defaults to the current date in
\code{YYYY-MM-DD} format.}

\item{silent}{Do not report progress. Defaults to \code{FALSE}.}

\item{overwrite}{Always overwrite existing files? Helpful for non-interactive use. Defaults to
\code{FALSE}.}

\item{as_is}{Import the schemas into the database without tidying? Defaults to \code{FALSE}.}
}
\value{
A database connection object of class
\link[RSQLite:SQLiteConnection-class]{RSQLite::SQLiteConnection}.
}
\description{
\code{save_db()} saves a list of UK Biobank data schemas (or, if \code{as_is == TRUE}, any list of data
frames) to an SQLite database.
}
\details{
\code{save_db()} takes a list of UK Biobank schemas and saves them to an SQLite
database. If \code{!as_is}, they are tidied. Note that if the table structure has
changed (i.e. has been changed by UK Biobank), then the function may fail
partially or fully unless \code{as_is == TRUE}.
}
\examples{
\dontrun{
sch <- ukbschemas()
db <- save_db(sch, path = tempdir())
}

}
