% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_schema_db.R
\name{save_schemas_db}
\alias{save_schemas_db}
\title{Save a list of UK Biobank data schemas to an SQLite database}
\usage{
save_schemas_db(sch, file = "", path = ".", date_str = Sys.Date(),
  silent = !interactive(), overwrite = FALSE, as_is = FALSE)
}
\description{
Save a list of UK Biobank data schemas to an SQLite database
}
