% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-db.R
\name{load_db}
\alias{load_db}
\title{Load schemas from a ukbschemas SQLite database}
\usage{
load_db(db_file = NULL, db = NULL)
}
\arguments{
\item{db_file}{Full path to an SQLite database, or \code{NULL}.}

\item{db}{A (possibly disconnected) database connection, or \code{NULL}.}
}
\value{
A named list with elements of class \code{tbl_df}, containing the tables
from \code{db}.
}
\description{
\code{load_db()} is a simple front-end to \link{DBI} functions which can be used to
load tables from a ukbschemas database (or in fact, any database; see
\link[DBI:DBI-package]{DBI::DBI-package}).
}
\details{
\code{load_db()} will attempt to open an SQLite database from file \code{db_file} or
from connection \code{db}. It will throw errors if the file is not a valid
database or it is impossible to create a valid connection from \code{db}. Tables
are read with \link[DBI:dbReadTable]{DBI::dbReadTable} and converted to data frames. The database
connection is always closed before the function returns its result.
}
\examples{
\dontrun{
db <- ukbschemas_db(path = tempdir())
sch <- load_db(db = db)
}

}
