% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-schemas.R
\name{load_schemas}
\alias{load_schemas}
\title{Load schemas from a ukbschema database}
\usage{
load_schemas(file = NULL, db = NULL)
}
\arguments{
\item{file}{Full path to an SQLite database, or \code{NULL}.}

\item{db}{A (possibly disconnected) database connection, or \code{NULL}.}
}
\value{
A named list with elements of class \link[tibble:tbl_df]{tibble::tbl_df}, containing the
tables from \code{db}.
}
\description{
\code{load_schemas()} is a simple front-end to \link[DBI:DBI]{DBI::DBI} functions which can be
used to load tables from a ukbschema database.
}
\details{
\code{load_schemas()} will attempt to open \code{file} as an SQLite database,
or will attempt to open connection \code{db} (if closed) and will throw errors if
the file is not a valid databse or it is impossible to create a valid
connection from \code{db}. Tables are read with \link[DBI:dbReadTable]{DBI::dbReadTable} and converted
to data frames of class \code{tbl_df}. The database connection is \emph{always} closed
before the function returns its result.
}
