% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbschemas.R
\name{ukbschemas}
\alias{ukbschemas}
\title{Fetch the UK Biobank data schemas via the internet}
\usage{
ukbschemas(silent = !interactive(), as_is = FALSE, debug = FALSE)
}
\arguments{
\item{silent}{Do not report progress. Defaults to \code{FALSE}.}

\item{as_is}{Import the schemas into the database without tidying? Defaults
to \code{FALSE}.}

\item{debug}{Report debugging information (useful when e.g. the structure of
the schema files has changed). Defaults to \code{FALSE}.}
}
\value{
A list of objects of class \code{tbl_df} (see \link[tibble:tibble]{tibble::tibble}).
}
\description{
\code{ukbschemas()} loads the UK Biobank data schemas into a list
}
\details{
The UK Biobank data schemas are fetched via the internet. If
\code{!as_is}, they are tidied. Note that if the table structure has changed
(i.e. has been changed by UK Biobank), then the function may fail partially
or fully.
}
