% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbschemas.R
\name{ukbschemas}
\alias{ukbschemas}
\title{Load the UK Biobank data schemas from the Internet or other repository}
\usage{
ukbschemas(
  sch_id = SCHEMA_FILENAMES[["id"]],
  sch_path = file.path(tempdir(), "ukbschemas", "schemas"),
  as_is = FALSE,
  nThread = detectCores(),
  silent = !interactive()
)
}
\arguments{
\item{sch_id}{The id numbers of the schemas to download. Type
\code{SCHEMA_FILENAMES} for reference. Defaults to all.}

\item{sch_path}{The directory path of the schema download cache. Defaults to
/tmp/RtmpKCAHiF/ukbschemas/schemas.}

\item{as_is}{Import the schemas into the database without tidying? Defaults
to \code{FALSE}.}

\item{nThread}{Number of threads to spawn for parallelisable tasks, such as
the downloading and importing of the schemas. Defaults to the number of
logical cores present on the system.}

\item{silent}{Do not report progress. Defaults to \code{FALSE}.}
}
\value{
A list of objects of class \code{data.frame}.
}
\description{
Load the UK Biobank data schemas from the Internet or other repository
}
\details{
The UK Biobank data schemas are fetched via the Internet unless a  different
\code{url_prefix} is provided. If \code{!as_is}, they are tidied. Note that  if the
table structure has changed (i.e. has been changed by UK Biobank), then the
function may fail partially or fully unless \code{as_is == TRUE}.
}
\note{
This is a convenience function to load the schemas directly from the UK
Biobank website. For most purposes, it is recommended to use \link{ukbschemas_db}
and \link{load_db} instead.
}
\examples{
\dontrun{
sch <- ukbschemas()
}

}
